feat: signup submission hook have been implemented
