+++ frontend/src/components/PetAd/PetAd.css
@@ -11,22 +11,22 @@
     box-shadow: 4px 4px 5px 5px rgba(0, 0, 0, 0.2);
     position: relative;
     font-size: 1rem;
-    display: flex; /* Add flex display */
-    flex-direction: column; /* Ensure the children are in a column */
-    align-items: center; /* Center horizontally */
+    display: flex; 
+    flex-direction: column; 
+    align-items: center; 
 }
 
 .pet-inf {
     display: flex;
     flex-direction: column;
-    gap: 10px; /* Adjust the value to set the desired vertical gap */
-    width: 100%; /* Ensure the pet-inf takes full width of pet-ad */
+    gap: 10px; 
+    width: 100%; 
 }
 
 
 .pet-ad img {
-    max-width: 100%; /* Ensure the image fits within the parent container */
-    height: auto; /* Maintain aspect ratio */
+    max-width: 100%; 
+    height: auto; 
 }
 
 .content{
@@ -34,7 +34,7 @@
 }
 
 .pet-info {
-    font-family: "Quicksand", sans-serif; /* Use your desired font-family */
-    font-weight: 700; /* Use the desired font weight */
-    color: #333; /* Use the desired color */
+    font-family: "Quicksand", sans-serif; 
+    font-weight: 700; 
+    color: #333; 
 }
