+++ frontend/src/hooks/useSignUp.ts
@@ -1,0 +1,52 @@
+import { useState } from "react";
+import axios from "axios";
+
+const useSignUp = () => {
+    const [nome, setNome] = useState("");
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const [confirmPassword, setConfirmPassword] = useState("");
+    const [errorMessage, setErrorMessage] = useState("");
+    const [successMessage, setSuccessMessage] = useState("");
+
+    const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+        event.preventDefault();
+        if (password !== confirmPassword) {
+            setErrorMessage("As senhas não coincidem. Por favor, tente novamente.");
+            setSuccessMessage("");
+        } else {
+            setErrorMessage("");
+            try {
+                await axios.post("http://localhost:3333/users", {
+                    name: nome,
+                    email: email,
+                    password: password
+                });
+                setSuccessMessage("Cadastro realizado com sucesso!");
+                setNome("");
+                setEmail("");
+                setPassword("");
+                setConfirmPassword("");
+            } catch (error) {
+                setErrorMessage("Erro ao realizar cadastro. Por favor, tente novamente.");
+                setSuccessMessage("");
+            }
+        }
+    };
+
+    return {
+        nome,
+        setNome,
+        email,
+        setEmail,
+        password,
+        setPassword,
+        confirmPassword,
+        setConfirmPassword,
+        errorMessage,
+        successMessage,
+        handleSubmit
+    };
+};
+
+export default useSignUp;
