+++ frontend/src/pages/Login/index.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import DescriptionPage from "../../components/DescriptionTemplate";
 import './styles.css';
@@ -8,6 +9,8 @@ const Login: React.FC = () => {
     const [password, setPassword] = useState("");
     const [errorMessage, setErrorMessage] = useState("");
 
+    const navigate = useNavigate();
+
     const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         try {
@@ -16,7 +19,7 @@ const Login: React.FC = () => {
                 password: password
             });
             document.cookie = `jwt=${response.data.token}; path=/;`;
-            window.location.href = "/"; 
+            navigate("/"); // Redirecionar para a página inicial após o login        
         } catch (error) {
             setErrorMessage("Email ou senha inválidos. Por favor, tente novamente.");
         }
