+++ frontend/src/pages/SignUp/styles.css
@@ -65,7 +65,7 @@
 
 .signup-links {
     text-align: center;
-    margin-top: 20px;
+    margin-top: 7px;
 }
 
 .signup-links a {
@@ -86,5 +86,5 @@
 
 .message {
     text-align: center;
-    margin-top: 15px; 
+    margin-top: 7px; 
 }
